Refactor main.py and fix most of failing tests
